requirements.txt
fastmcp==2.12.4
pydantic==2.11.9
python-dotenv==1.1.1
mysql-connector-python==9.4.0
neo4j==5.28.2
uvicorn==0.35.0
sse-starlette==3.0.2